FIX/OPT: circular dep; doc; remove unused pkg
