+++ pyproject.toml
@@ -11,10 +11,8 @@ python = "^3.9"
 python-dotenv = "^1.0.0"
 docker = "^6.1.3"
 requests = "^2.31.0"
-matplotlib = "^3.7.1"
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
-gevent = "^22.10.2"
 locust = "^2.15.1"
 pyyaml = "^6.0"
 ping3 = "^4.0.4"
@@ -23,6 +21,7 @@ urllib3 = "<2.0"
 asciichartpy = "^1.5.25"
 numpy = "^1.25.1"
 astropy = "^5.3.1"
+python-dateutil = "^2.8.2"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
