+++ src/collect.py
@@ -1,19 +1,30 @@
+import docker
+from ping3 import ping
+
 import argparse
 import sys
 import os
 import time
-from typing import List
 import subprocess
 import shutil
-from src.type import CollectCommandArg, State, ChangedRule
-from src.utils import UtilMapper, container_is_healthy, logger
-from ping3 import ping
+from typing import List
 
+from src.model import CollectCommandArg, UtilMapper
+from src.type import State
+from src.utils import logger
+
+
+class _ChangedRule:
+    # @TODO: implement id
+    req: str
+
+    def __init__(self, req: str):
+        self.req = req
 
 def get_test_command_arg(*args) -> CollectCommandArg:
-    parser = argparse.ArgumentParser(description='WAF Test Command Parser')
+    parser = argparse.ArgumentParser(description='WAF Collect Command Parser')
     parser.add_argument('--test-name', type=str, help='test name', required=True)
-    parser.add_argument('--utils', type=str, help='utils', required=True)
+    parser.add_argument('--utils', type=str, help='utils')
     parser.add_argument('--before', type=str, help='before test command')
     parser.add_argument('--after', type=str, help='after test command')
     parser.add_argument('--raw-output', type=str, help='raw output')
@@ -26,7 +37,7 @@ def get_test_command_arg(*args) -> CollectCommandArg:
         test_name=parsed_args.test_name,
         before=parsed_args.before,
         after=parsed_args.after,
-        utils=parsed_args.utils.split(","),
+        utils=None if parsed_args.utils is None else parsed_args.utils.split(","),
         raw_output=parsed_args.raw_output,
         output=parsed_args.output,
         waf_endpoint=parsed_args.waf_endpoint
@@ -50,21 +61,21 @@ def init(arg: CollectCommandArg):
     os.makedirs(arg.after_rules_dir, exist_ok=True)
     os.makedirs(arg.test_cases_dir, exist_ok=True)
 
-def get_changed_rules(arg: CollectCommandArg) -> List[ChangedRule]:
+def get_changed_rules(arg: CollectCommandArg) -> List[_ChangedRule]:
     exec_cmd = f"git diff --name-only {arg.before} {arg.after} | grep -E \'rules/.*.conf$\'"
     
     subprocess.run(exec_cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     output = subprocess.check_output(exec_cmd, shell=True).decode()
         
-    changedRules: List[ChangedRule] = []
+    changedRules: List[_ChangedRule] = []
         
     for line in output.splitlines():
         logger.info(f"Changed rule: {line}")
-        changedRules.append(ChangedRule(line.replace("rules/", "").replace(".conf", "")))
+        changedRules.append(_ChangedRule(line.replace("rules/", "").replace(".conf", "")))
 
     return changedRules
 
-def init_tmp_file(arg: CollectCommandArg, changedRules: List[ChangedRule]):
+def init_tmp_file(arg: CollectCommandArg, changedRules: List[_ChangedRule]):
     # copy before-rules
     for changedRule in changedRules:
         cmd = f"git show {arg.before}:rules/{changedRule.req}.conf > {arg.before_rules_dir}/{changedRule.req}.conf"
@@ -92,13 +103,14 @@ def init_tmp_file(arg: CollectCommandArg, changedRules: List[ChangedRule]):
 def init_docker_compose_file(arg: CollectCommandArg, state: State):
     shutil.copyfile("./tests/docker-compose.yml", f"./tests/docker-compose-{state.name}.yml")
     processed_path = (arg.before_rules_dir if state == State.before else arg.after_rules_dir).replace("/", "\\/")
-    cmd = f"""sed -i -e "s/- ..\/rules/- ..\/{processed_path}/g" .\/tests\/docker-compose-{state.name}.yml"""
+    cmd = f"""sed -i -e "s/- ..\\/rules/- ..\\/{processed_path}/g" .\\/tests\\/docker-compose-{state.name}.yml"""
+    # cmd = f"""sed -i -e "s/- ..\/rules/- ..\/{processed_path}/g" .\/tests\/docker-compose-{state.name}.yml"""
     subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
     # @TODO: this is a temporary fix for extra files created from sed
     os.remove(f"./tests/docker-compose-{state.name}.yml-e")
 
-def runner(args: CollectCommandArg, changedRules: List[ChangedRule], state: State):
+def runner(args: CollectCommandArg, state: State):
     # start service with docker-compose
     cmd = f"docker-compose -f ./tests/docker-compose-{state.name}.yml up -d {args.modsec_version}"
 
@@ -126,7 +138,7 @@ def waf_server_is_up(waf_endpoint: str) -> bool:
     
     # @TODO: update
     while retry < 3:
-        if container_is_healthy("modsec2-apache"):
+        if docker.from_env().api.inspect_container("modsec2-apache")["State"]["Status"] == 'running':
             return True
         retry += 1
         time.sleep(timeout)
@@ -167,7 +179,7 @@ def main():
     init_tmp_file(args, changedRules)
     
     init_docker_compose_file(args, State.before)
-    runner(args, changedRules, State.before)
+    runner(args, State.before)
     
     init_docker_compose_file(args, State.after)
-    runner(args, changedRules, State.after)
+    runner(args, State.after)
