+++ src/model/CAdvisorUtil.py
@@ -1,16 +1,15 @@
 import docker
+
 import requests
 import time
 import subprocess
 import json
 import os
-from typing import List
-from src.model.Util import Util
-from src.model.Threshold import Threshold
-from src.model.ParsedDataItem import ParsedDataItem
-from src.type import CollectCommandArg, State, ReportCommandArg
-from .fn import container_is_healthy, create_time_series_terminal_plot, save_json
-from .logger import logger
+from typing import Type, List
+
+from .Util import Util, ParsedDataItem, CollectCommandArg, ReportCommandArg
+from src.type import State
+from src.utils import logger
 
 
 class CAdvisorUtil(Util):
@@ -19,7 +18,9 @@ class CAdvisorUtil(Util):
     """
     
     # @TODO: add to variables
+    __cAdvisor_endpoint: str = "http://127.0.0.1:8080/api/v1.1/subcontainers/docker/"
     __waf_container_name: str = "modsec2-apache"
+    __cAdvisor_container_version: str = "v0.45.0"
     raw_filename: str = "cAdvisor.json"
     
     def collect(self, args: CollectCommandArg, state: State = None):
@@ -35,7 +36,7 @@ class CAdvisorUtil(Util):
             )
 
         data_list, timestamp_set = [], set()
-        url = f"http://127.0.0.1:8080/api/v1.1/subcontainers/docker/{self.__get_waf_container_id()}"
+        url = f"{self.__cAdvisor_endpoint}{self.__get_waf_container_id()}"
 
         # cAdvisor API sends 60 recent dataset,
         # so the data requires to be filtered out duplicates
@@ -44,7 +45,7 @@ class CAdvisorUtil(Util):
             time.sleep(10)
         
         self.fetch_data(data_list, timestamp_set, url)
-        save_json(f"{args.raw_output}/{state.name}_{self.raw_filename}", data_list)
+        self.save_json(f"{args.raw_output}/{state.name}_{self.raw_filename}", data_list)
         self.__stop_cadvisor()
     
     def text_report(self, args: ReportCommandArg):
@@ -52,16 +53,17 @@ class CAdvisorUtil(Util):
         after_data = self.parse_data(f"{args.raw_output}/{State.after.name}_{self.raw_filename}")
         
         for matrix in ["cpu_total", "cpu_user", "cpu_system", "memory_usage", "memory_cache"]:
-            print(create_time_series_terminal_plot(matrix, before_data[matrix], after_data[matrix]))
+            print(self.create_time_series_terminal_plot(matrix, before_data[matrix], after_data[matrix]))
             
         if not args.threshold_conf:
             return
         
-        thresholds: List[Threshold] = self._get_threshold(os.path.join(args.threshold_conf, "cAdvisor.threshold.json"))
+        thresholds = self._get_threshold(os.path.join(args.threshold_conf, "cAdvisor.threshold.json"))
 
         for threshold in thresholds:
             threshold.inspect(before_data[threshold.metric_name], after_data[threshold.metric_name])
 
+    # @TODO: impl
     def figure_report(self, args: ReportCommandArg):
         pass
 
@@ -117,26 +119,27 @@ class CAdvisorUtil(Util):
             exit(1)
     
     def __start_cadvisor(self):
-        cmd = """
-        docker run \
-        --volume=/:/rootfs:ro \
-        --volume=/var/run:/var/run:ro \
-        --volume=/sys:/sys:ro \
-        --volume=/var/lib/docker/:/var/lib/docker:ro \
-        --volume=/dev/disk/:/dev/disk:ro \
-        --publish=8080:8080 \
-        --detach=true \
-        --name=cadvisor \
-        --privileged \
-        --device=/dev/kmsg \
-        gcr.io/cadvisor/cadvisor:v0.45.0
-        """    
-        
+
         try:
+            cmd = f"""
+            docker run \
+            --volume=/:/rootfs:ro \
+            --volume=/var/run:/var/run:ro \
+            --volume=/sys:/sys:ro \
+            --volume=/var/lib/docker/:/var/lib/docker:ro \
+            --volume=/dev/disk/:/dev/disk:ro \
+            --publish=8080:8080 \
+            --detach=true \
+            --name=cadvisor \
+            --privileged \
+            --device=/dev/kmsg \
+            gcr.io/cadvisor/cadvisor:{self.__cAdvisor_container_version}
+            """
+            
             subprocess.run(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
             logger.info("Waiting for cAdvisor to be up...")
             cnt = 0
-            while not container_is_healthy("cadvisor") and cnt < 6:
+            while not self.container_is_healthy("cadvisor") and cnt < 6:
                 time.sleep(10)
                 cnt += 1
             time.sleep(30)
@@ -145,9 +148,32 @@ class CAdvisorUtil(Util):
             exit(1)
     
     def __stop_cadvisor(self):
+        """
+        stop the cAdvisor container
+        """
         cmd = "docker stop cadvisor && docker rm cadvisor"
         try:
             subprocess.run(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
         except Exception as e:
             logger.error(e)
             exit(1)
+            
+    def save_json(self, dist_path: str, data: any, cls: Type[json.JSONEncoder] = None):
+        """_summary_
+
+        Args:
+            dist_path (str): _description_
+            data (any): _description_
+            cls (Type[json.JSONEncoder], optional): _description_. Defaults to None.
+
+        Returns:
+            _type_: _description_
+        """
+
+        os.makedirs(os.path.dirname(dist_path), exist_ok=True)
+        with open(dist_path, "w+") as file:
+            json.dump(data, file, indent=2, cls=cls)
+        file.close()
+    
+    def container_is_healthy(self, name_or_id: str) -> bool:
+        return docker.from_env().api.inspect_container(name_or_id)["State"]["Status"] == 'running'
