+++ src/model/FTWUtil.py
@@ -3,10 +3,9 @@ import os
 import json
 import time
 from typing import List
-from src.utils import logger, color_text, create_colored_text_by_value
-from src.model.Util import Util, Threshold
-from src.model.ParsedDataItem import ParsedDataItem
-from src.type import CollectCommandArg, State, ReportCommandArg
+
+from .Util import ParsedDataItem, Util, ReportCommandArg, CollectCommandArg
+from src.type import State
 
 
 REPORT_PLAIN_TEXT_FORMAT: str = (
@@ -40,16 +39,16 @@ class FTWUtil(Util):
         time.sleep(5)
         
         command = f'go-ftw run -d {args.test_cases_dir} -o json > {args.raw_output}/{state.name}_{self.raw_filename}'
-        _ = subprocess.run(command, shell=True, check=False, stderr=subprocess.DEVNULL, stdout=subprocess.DEVNULL)
+        subprocess.run(command, shell=True, check=False, stderr=subprocess.DEVNULL, stdout=subprocess.DEVNULL)
     
     def text_report(self, args: ReportCommandArg):
         before_data = self.parse_data(f"{args.raw_output}/{State.before.name}_{self.raw_filename}")
         after_data = self.parse_data(f"{args.raw_output}/{State.after.name}_{self.raw_filename}")
 
-        run_changes = create_colored_text_by_value(before_data["run"].value - after_data["run"].value)
-        success_changes = create_colored_text_by_value(len(before_data["success"]) - len(after_data["success"]))
-        failed_changes = create_colored_text_by_value(len(before_data["failed"]) - len(after_data["failed"]))
-        total_time_changes = create_colored_text_by_value(before_data["totalTime"].value - after_data["totalTime"].value)
+        run_changes = self.create_colored_text_by_value(before_data["run"].value - after_data["run"].value)
+        success_changes = self.create_colored_text_by_value(len(before_data["success"]) - len(after_data["success"]))
+        failed_changes = self.create_colored_text_by_value(len(before_data["failed"]) - len(after_data["failed"]))
+        total_time_changes = self.create_colored_text_by_value(before_data["totalTime"].value - after_data["totalTime"].value)
         
         # generate report
         report = REPORT_PLAIN_TEXT_FORMAT.format(
@@ -73,7 +72,7 @@ class FTWUtil(Util):
         if not args.threshold_conf:
             return
 
-        thresholds: List[Threshold] = self._get_threshold(os.path.join(args.threshold_conf, "ftw.threshold.json"))
+        thresholds = self._get_threshold(os.path.join(args.threshold_conf, "ftw.threshold.json"))
 
         for threshold in thresholds:
             threshold.inspect(before_data[threshold.metric_name], after_data[threshold.metric_name])
