--- src/type/ChangedRule.py
@@ -1,8 +1,0 @@
-class ChangedRule:
-    # @TODO: implement id
-    id: str
-    req: str
-
-    def __init__(self, req: str):
-        self.id = None
-        self.req = req
