+++ src/utils/__init__.py
@@ -1,29 +1,4 @@
-from .fn import *
-from .logger import *
-from .cAdvisor import *
-from .ftw import *
-from .locust import *
-from src.type import UtilType
+from .logger import logger
 
 
-UtilMapper: dict[str, Util] = {
-    UtilType.ftw.name: FTWUtil,
-    UtilType.cAdvisor.name: CAdvisorUtil,
-    UtilType.locust.name: LocustUtil
-}
-
-__all__ = [
-    "CAdvisorUtil",
-    "FTWUtil",
-    "LocustUtil",
-    "logger",
-    "container_is_healthy",
-    "color_text",
-    "get_terminal_column",
-    "create_time_series_terminal_plot",
-    "iso_time_str_to_unix_time",
-    "UtilMapper",
-    "save_json",
-    "create_data_diff_terminal_table",
-    "create_colored_text_by_value"
-]
+__all__ = ["logger"]
